ati-client/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ati-client"
version = "0.2.0"
description = "Python SDK for ATI (Agent Tools Interface) — token utilities and orchestrator provisioning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Parcha", email = "eng@parcha.com" },
]
keywords = ["ati", "agent", "tools", "mcp", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://github.com/Parcha-ai/ati"
Repository = "https://github.com/Parcha-ai/ati"
Issues = "https://github.com/Parcha-ai/ati/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ati"]

[tool.pytest.ini_options]
testpaths = ["tests"]
